+++ Modules/Intent.Modules.Eventing.MassTransit/Api/IRabbitMQConsumerSettings.cs
@@ -1,8 +1,10 @@
 namespace Intent.Eventing.MassTransit.Api;
 
+// This interface will allow us to seamlessly use any "RabbitMQConsumerSettings" stereotype that's been assigned to various elements.
 public interface IRabbitMQConsumerSettings
 {
     string Name { get; }
+    string EndpointName();
     int? PrefetchCount();
     bool Lazy();
     bool Durable();