+++ Modules/Intent.Modules.Eventing.MassTransit.RequestResponse/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.1.0
+
+- Improvement: Updated according to changes made in `Intent.Eventing.MassTransit` module.
+
 ### Version 1.0.0
 
 - New Feature: Add `Message Triggered` stereotypes to `Command`/`Query` elements to expose them as Consumers for MassTransit. Setup Service Proxies on the client side to interact with those endpoints as though they were HTTP based.